refactor(tests): rename tests/fakes.py -> tests/_broker_fakes.py
Match the repo's established shared-fakes convention: fakes live in
tests/_<domain>_fakes.py (underscore-prefixed, domain-scoped) alongside
_fanout_fakes.py / _peer_fakes.py / _capability_fakes.py / etc. Bare
tests/fakes.py was the odd one out. Import-only change across the 9
consumers; no behavior change.
